+++ docs/changelog.rst
@@ -5,6 +5,8 @@ In Development
 ++++++++++++++++++++++++++
 * Support `include_missing_properties` bravado-core flag in pyramid configuration
 * Outsource flattening logic to bravado-core library.
+* Expose bravado-core ``operation`` in request object
+* Add ``pyramid_renderer`` and ``PyramidSwaggerRendererFactory``
 
 2.4.1 (2017-06-14)
 ++++++++++++++++++++++++++